+++ wayt_app/lib/repositories/travel_item_repository/repository/repository.dart
@@ -1,2 +1,2 @@
-export 'events/events.dart';
 export 'travel_item_repository.dart';
+export 'travel_item_repository_states.dart';
